Change modules to incomplete/library status
This at least got the build to work; also eliminate exceptions being thrown in constructors which Dagger cannot seem to handle (switch to log output instead). Also switch from Lazy<> loads to just normal @Inject for the example to pull in the library dependencies which seems to work fine.

All of this needs to be tested now...