Make endpoint URI's relative to their controller context
This way a controller at context "/store" with endpoint "/fruit" becomes accessible as "/store/fruit"; later, if we change the controller's base context to "/shop", then "/fruit" carries with it and remaps to "/shop/fruit"; this allows us to hierarchically construct realms of responsibility for controllers and endpoints.

There are a couple caveats: this removes the regex from the mapping call, always causing the mapping to take /store/fruit*** (eveyrthing that starts with this). If we wanted to have a more sophisticated mapping, currently we would not be able to support it. We can refactor later to account for this as needed.